Fixed  book delete errors

+++ controllers/bookControllers.js
@@ -1,4 +1,5 @@
 const Book = require("../models/Book")
+const mongoose = require("mongoose")
 
 exports.createBook = async (req, res) => {
   try {
@@ -56,7 +57,10 @@ exports.updateBook = async (req, res) => {
 
 exports.deleteBook = async (req,res) => {
    try {
-    const id = req.params
+    const {id} = req.params
+   if(!mongoose.Types.ObjectId.isValid(id)){
+     return res.status(400).json({ message: "Invalid book ID" })
+   }
     const book = await Book.findByIdAndDelete(id)
     if(!book){
         return res.status(404).json({"message":"Book not found"})
